Fix PipeWire compatibility issue.
Under PipeWire, when the card profile is changed, existing sink inputs
are migrated to a new sink named after the card and the new profile.

This causes flapping, if the determination of whether a card has
changed from/to multichannel being needed, is based on the sink.

We therefore aggregate sink inputs by card, not by sink, ensuring
the flapping cannot take place anymore.

Also use a dynamic timeout to avoid using too many CPU resources.
Furthermore, to reduce CPU usage anyway, subscribe only to the
events needed to provide the functionality this program provides.